+++ geminidr/core/parameters_nearIR.py
@@ -51,11 +51,17 @@ class cleanReadoutConfig(config.Config):
     debug_subtract_background = config.Field("Subtract median from each pattern box?", bool, True)
     level_bias_offset = config.Field("Level the bias offset across (sub-)quads accompanying pattern noise?", bool, True)
     smoothing_extent = config.RangeField("Width (in pix) of the region at a given quad interface to be smoothed over", int, 5, min=1)
+    intraquad_smooth = config.RangeField("Height (in pix) of the region at a bias jump to be smoothed over", int, 50, min=1)
     sg_win_size = config.RangeField("Smoothing window size (pixels) for Savitzky-Golay filter", int, 25, min=3)
     simple_thres = config.RangeField("Pattern edge detection threshold", float, 0.6, min=0.0)
     pat_strength_thres = config.RangeField("Pattern strength threshold", float, 15.0, min=0.0)
-    clean = config.Field("Behavior of the routine? Must be one of default, skip, or force", str, "default")
+    clean = config.ChoiceField("Cleaning behavior", str,
+                               allowed={"default": "perform pattern removal if pattern in strong enough",
+                                        "force": "force pattern removal",
+                                        "skip": "skip primitive"},
+                               default="default", optional=False)
     debug_canny_sigma = config.RangeField("Standard deviation for smoothing of Canny edge-finding", float, 3, min=1)
+
 
 class cleanFFTReadoutConfig(config.Config):
     suffix = config.Field("Filename suffix", str, "_readoutFFTCleaned", optional=True)
@@ -66,8 +72,12 @@ class cleanFFTReadoutConfig(config.Config):
     lquad = config.Field("Level the bias offset across (sub-)quads accompanying pattern noise?", bool, True)
     l2clean = config.Field("Clean Fourier artifacts?", bool, True)
     l2thres = config.RangeField("Sigma factor to be used in thresholding for l2clean", float, 4., min=0)
-    clean = config.Field("Behavior of the routine? Must be one of default, skip, or force", str, "default")
+    clean = config.ChoiceField("Cleaning behavior", str,
+                               allowed={"default": "perform pattern removal if pattern in strong enough",
+                                        "skip": "skip primitive"},
+                               default="skip", optional=False)
     smoothing_extent = config.RangeField("Width (in pix) of the region at a given quad interface to be smoothed over", int, 5, min=1)
+
 
 class separateFlatsDarksConfig(config.Config):
     pass