+++ geminidr/core/tests/test_nearIR.py
@@ -1,6 +1,7 @@
 """
 Tests applied to primitives_nearIR.py
 """
+import pytest
 
 from datetime import datetime
 import os
@@ -9,7 +10,7 @@ import astrodata
 from astrodata.testing import ad_compare
 from geminidr.core.tests.test_spect import create_zero_filled_fake_astrodata
 from geminidr.core import primitives_nearIR
-import pytest
+from recipe_system.mappers.primitiveMapper import PrimitiveMapper
 
 
 # ad input list maker functions
@@ -100,6 +101,7 @@ def test_remove_first_frame_by_filename():
 @pytest.mark.preprocessed_data
 @pytest.mark.parametrize("in_file",
                          ["S20060826S0305",  # GNIRS LS (parameters need tweaks)
+                          "N20231112S0137",  # GNIRS LS (spectrum across boundary)
                           "N20050614S0190",  # NIRI LS (parameters need tweaks)
                           "N20170505S0146",  # NIRI image, single star
                           "N20220902S0145",  # NIRI image, extended source
@@ -114,11 +116,15 @@ def test_clean_readout(in_file, path_to_inputs, path_to_refs):
     ad = astrodata.open(os.path.join(path_to_inputs,
                                      in_file + '_skyCorrected.fits'))
 
-    p = primitives_nearIR.NearIR([ad])
+    # Must use the correct default parameters, since this is a test that the
+    # defaults haven't changed
+    pm = PrimitiveMapper(ad.tags, ad.instrument(generic=True).lower(),
+                         mode="sq", drpkg="geminidr")
+    pclass = pm.get_applicable_primitives()
+    p = pclass([ad])
     ad_out = p.cleanReadout(clean="default")[0]
 
-    ref = astrodata.open(os.path.join(path_to_refs,
-                                      in_file + '_readoutCleaned.fits'))
+    ref = astrodata.open(os.path.join(path_to_refs, ad.filename))
     assert ad_compare(ad_out, ref)
 
 
@@ -127,15 +133,21 @@ def test_clean_readout(in_file, path_to_inputs, path_to_refs):
 @pytest.mark.parametrize("in_file",
                          ["N20220902S0145",  # NIRI image, extended source
                           "N20101227S0040",  # GNIRS LS (par needs tweaking pat_thres=0.1). Only FFT can handle this frame.
-                          "S20060826S0305",
-                          "N20170505S0146",
-                          "N20060103S0010",
-                          "N20060218S0138",
-                          "N20051120S0378",
+                          "S20060826S0305",  # GNIRS LS
+                          "N20170505S0146",  # NIRI image
+                          "N20060103S0010",  # NIRI image
+                          "N20060218S0138",  # NIRI image
+                          "N20051120S0378",  # NIRI image
+                          "N20231112S0136",  # GNIRS LS
                           ])
 def test_clean_fftreadout(in_file, path_to_inputs, path_to_refs):
     ad = astrodata.open(os.path.join(path_to_inputs, in_file + '_skyCorrected.fits'))
-    p = primitives_nearIR.NearIR([ad])
+    # Must use the correct default parameters, since this is a test that the
+    # defaults haven't changed
+    pm = PrimitiveMapper(ad.tags, ad.instrument(generic=True).lower(),
+                         mode="sq", drpkg="geminidr")
+    pclass = pm.get_applicable_primitives()
+    p = pclass([ad])    
     ad_out = p.cleanFFTReadout(clean="default")[0]
     ref = astrodata.open(os.path.join(path_to_refs, in_file + '_readoutFFTCleaned.fits'))
     assert ad_compare(ad_out, ref)