+++ geminidr/gmos/tests/spect/test_distortion_correct_with_wavelength_solution.py
@@ -96,7 +96,7 @@ def test_regression_in_distortion_correct_with_wave_sol(ad, change_working_dir, 
 
         p = primitives_gmos_longslit.GMOSLongslit([deepcopy(ad)])
         p.viewer = geminidr.dormantViewer(p, None)
-        p.distortionCorrect(order=3, subsample=1)
+        p.distortionCorrect(interpolant="spline3", subsample=1)
         dist_corrected_ad = p.writeOutputs()[0]
 
     ref_ad = ref_ad_factory(dist_corrected_ad.filename)