+++ geminidr/niri/parameters_niri_longslit.py
@@ -1,2 +1,15 @@
 # This parameter file contains the parameters related to the primitive located
 # in the primitives_niri_longslit.py file, in alphabetical order.
+from geminidr.core import parameters_nearIR
+
+
+class cleanReadoutConfig(parameters_nearIR.cleanReadoutConfig):
+    # Need a larger extent to cope with a bright spectrum down the middle
+    def setDefaults(self):
+        self.smoothing_extent = 300
+
+
+class cleanFFTReadoutConfig(parameters_nearIR.cleanFFTReadoutConfig):
+    # Need a larger extent to cope with a bright spectrum down the middle
+    def setDefaults(self):
+        self.smoothing_extent = 300